lunchbreak/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file lunchbreak/requirements.txt lunchbreak/requirements.in
#
-e git+git://github.com/AndreasBackx/django-jinja.git@error-view-extensions#egg=django-jinja
-e git+git://github.com/RacingTadpole/django-private-media.git@7510f2f63ddf0653679b4134a0542cd78317a5c8#egg=django_private_media
-e git+git://github.com/AndreasBackx/django-push-notifications.git@269816a469c93dfcc8e87e910541c9f517aa540f#egg=django_push_notifications
-e git+git://github.com/AndreasBackx/google-maps-services-python.git@master#egg=googlemaps
-e git+git://github.com/AndreasBackx/pendulum@feature/period-exclude-merge#egg=pendulum
amqp==2.1.4               # via kombu
asn1crypto==0.22.0        # via cryptography
bcrypt==3.1.3
billiard==3.5.0.2         # via celery
celery==4.0.2
cffi==1.10.0              # via bcrypt, cryptography
contextlib2==0.5.5        # via raven
cryptography==1.8.2
django-appconf==1.0.2     # via django-compressor, django-imagekit
django-compressor==2.1.1
django-dirtyfields==1.1
django-imagekit==3.3
django-jsonfield==1.0.1
django-phonenumber-field==0.7.2
django-safedelete==0.4.0
django-sass-processor==0.3.5
django-sendfile==0.3.11
django-split-settings==0.1.2
django-subdomains==2.1.0
Django==1.9.12            # via django-dirtyfields, django-sendfile, django-split-settings, django-subdomains
djangorestframework-jsonp==1.0.0
djangorestframework==3.4.6
drf-extensions==0.2.8
et-xmlfile==1.0.1         # via openpyxl
gocardless-pro==0.2.0
idna==2.5                 # via cryptography, twilio
jdcal==1.3                # via openpyxl
Jinja2==2.8
kombu==4.0.2              # via celery
libsass==0.13.2           # via django-sass-processor
MarkupSafe==1.0           # via jinja2
mysqlclient==1.3.10
openpyxl==2.3.4
packaging==16.8           # via cryptography
phonenumbers==8.5.1       # via django-phonenumber-field
pilkit==2.0               # via django-imagekit
Pillow==3.2.0
plivo==0.11.3
psycopg2==2.7.1
pycparser==2.17           # via cffi
PyJWT==1.5.0              # via twilio
pyOpenSSL==17.0.0         # via twilio
pyparsing==2.2.0          # via packaging
pysocks==1.6.7            # via twilio
python-dateutil==2.6.0
pytz==2017.2              # via celery, django-dirtyfields, twilio, tzlocal
pyyaml==3.11
raven==6.0.0
rcssmin==1.0.6            # via django-compressor
requests==2.9.1
rjsmin==1.0.12            # via django-compressor
six==1.10.0               # via bcrypt, cryptography, django-imagekit, libsass, packaging, pyopenssl, python-dateutil, twilio
twilio==6.1.0
tzlocal==1.4
ua-parser==0.7.1
user-agents==1.0.1
vine==1.1.3               # via amqp

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via cryptography
